refactor(tests): co-locate test files with source modules
Move test files from flat tests/ directory to live alongside the source
they test. Update jest.config.js, tsconfig.json, and package.json globs
accordingly. Closes #37.
