+++ jest.config.js
@@ -2,6 +2,6 @@
 module.exports = {
   preset: "ts-jest",
   testEnvironment: "node",
-  testMatch: ["**/tests/**/*.test.ts"],
+  testMatch: ["**/*.test.ts"],
   moduleFileExtensions: ["ts", "js", "json"],
 };
